requirements.txt
# Core Django
Django==4.2.7
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
django-filter==23.3
drf-yasg==1.21.7

# Authentication & Authorization
django-allauth==0.57.0
dj-rest-auth==5.0.2
PyJWT==2.8.0

# UI & Admin
django-jazzmin==2.6.0
django-bootstrap4==23.2

# AI & Document Processing
openai==1.12.0
python-docx==1.1.0
httpx==0.27.2

# Storage & AWS
django-storages==1.14.2
boto3==1.34.0

# Background Tasks & Caching
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
redis==5.0.1

# CORS
django-cors-headers==4.3.1

# Static Files
whitenoise==6.6.0

# Backup
django-dbbackup==4.0.2

# Cron Jobs
django-crontab==0.7.1

# Development Tools
django-extensions==3.2.3

# Payment Processing
stripe==7.8.0

# OTP Authentication
pyotp==2.9.0

# Server
gunicorn==21.2.0

# Testing (Development)
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0

# Code Quality (Development)
black==23.11.0


pillow==11.3.0

flower

sendgrid
